requirements.txt
Flask==2.3.2
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0